DEFAULT_LOAD_FACTOR was changed

+++ src/main/java/core/basesyntax/MyHashMap.java
@@ -3,7 +3,7 @@ package core.basesyntax;
 import java.util.Objects;
 
 public class MyHashMap<K, V> implements MyMap<K, V> {
-    private static final double DEFAULT_LOAD_FACTOR = 0.75f;
+    private static final float DEFAULT_LOAD_FACTOR = 0.75f;
     private static final int GROWING_COEFICIENT = 2;
     private static final int DEFAULT_CAPACITY = 16;
     private Node<K, V>[] table;
